+++ NEWS.md
@@ -5,6 +5,7 @@
 * Add option `rarefy_nb_seqs` to `ggven_pq()` in order to rarefy samples before plotting. 
 * Add function `SRS_curve_pq()` to plot scaling with ranked subsampling (SRS) curves using the `SRS::SRS_curve()` function (see citation("SRS") for reference).
 * Add option `nb_samples_info` to `biplot_pq()` in order to add the number of samples merged by level of factors.
+* Add a message when two modalities differ greatly (more than x2) in their number of sequences in `biplot_pq()` and `ggvenn_pq()`
 
 
 # MiscMetabar 0.33